# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mausam 2.0 - Personalized Weather Intelligence",
  "description": "Mausam 2.0 - a Personalized Weather Intelligence Engine on official IMD/MoES data. Personalized homepage cards for the 'Mausam' app (SIH 2026 PS 26076).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/admin/login": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Login",
    "operationId": "login_api_admin_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/app__admin__router__LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/admin/me": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Me",
    "operationId": "me_api_admin_me_get",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/overview": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Overview",
    "operationId": "overview_api_admin_overview_get",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_admin_users_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Get User",
    "operationId": "get_user_api_admin_users__user_id__get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Update User",
    "operationId": "update_user_api_admin_users__user_id__put",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Delete User",
    "operationId": "delete_user_api_admin_users__user_id__delete",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}/activities": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Activities",
    "operationId": "list_activities_api_admin_users__user_id__activities_get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}/activities/{activity_id}": {
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Delete Activity",
    "operationId": "delete_activity_api_admin_users__user_id__activities__activity_id__delete",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "activity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Activity Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/alerts": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Alerts",
    "operationId": "list_alerts_api_admin_alerts_get",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create Alert",
    "operationId": "create_alert_api_admin_alerts_post",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertCreate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/alerts/{alert_id}": {
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Update Alert",
    "operationId": "update_alert_api_admin_alerts__alert_id__put",
    "parameters": [
     {
      "name": "alert_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Alert Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Delete Alert",
    "operationId": "delete_alert_api_admin_alerts__alert_id__delete",
    "parameters": [
     {
      "name": "alert_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Alert Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/alerts/{alert_id}/impact": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Alert Impact",
    "operationId": "alert_impact_api_admin_alerts__alert_id__impact_get",
    "parameters": [
     {
      "name": "alert_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Alert Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/push/broadcast": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Push Broadcast",
    "description": "Send a push notification to every registered device (operator demo).",
    "operationId": "push_broadcast_api_admin_push_broadcast_post",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PushBroadcast"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AlertCreate": {
    "properties": {
     "severity": {
      "type": "string",
      "title": "Severity",
      "default": "orange"
     },
     "event_type": {
      "type": "string",
      "title": "Event Type",
      "default": "heavy_rain"
     },
     "headline": {
      "type": "string",
      "title": "Headline",
      "default": "Heavy rain expected"
     },
     "detail": {
      "type": "string",
      "title": "Detail",
      "default": ""
     },
     "region": {
      "type": "string",
      "title": "Region",
      "default": "Pune"
     },
     "impact": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Impact",
      "default": []
     },
     "valid_hours": {
      "type": "integer",
      "title": "Valid Hours",
      "default": 24
     },
     "radius_km": {
      "type": "number",
      "title": "Radius Km",
      "default": 40.0
     }
    },
    "type": "object",
    "title": "AlertCreate"
   },
...
   "AlertUpdate": {
    "properties": {
     "severity": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Severity"
     },
     "event_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Event Type"
     },
     "headline": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Headline"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail"
     },
     "region": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Region"
     },
     "valid_hours": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Valid Hours"
     }
    },
    "type": "object",
    "title": "AlertUpdate"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PushBroadcast": {
    "properties": {
     "title": {
      "type": "string",
      "title": "Title",
      "default": "Mausam"
     },
     "body": {
      "type": "string",
      "title": "Body",
      "default": ""
     },
     "region": {
      "type": "string",
      "title": "Region",
      "default": "All India"
     },
     "severity": {
      "type": "string",
      "title": "Severity",
      "default": "yellow"
     }
    },
    "type": "object",
    "title": "PushBroadcast"
   },
...
   "UserUpdate": {
    "properties": {
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "personas": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Personas"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "health": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Health"
     }
    },
    "type": "object",
    "title": "UserUpdate"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "app__admin__router__LoginRequest": {
    "properties": {
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "LoginRequest"
   },
...
  }
 }
}